Refactor Flask Routes
