+++ journal/weekX.md
@@ -85,5 +85,8 @@ npm run build
 ### Refactor JWT to use a decorator
 - Made substantial changes to implement a decorator. Primarily [`backend-flask/lib/cognito_jwt_token.py`](https://github.com/kmb40/aws-bootcamp-cruddur-2023/commit/f0e687796ccff97ab608fd8ac00cc66a219d11f2#diff-c4205e0c1fabfa8932a580cf9b109f3af4c9fe9c738868a9c7f8d83212db9c24) and [`backend-flask/app.py`](https://github.com/kmb40/aws-bootcamp-cruddur-2023/commit/6bc499cd9f426750930c4ede9bd21c075be52a43?diff=unified).
 
-### Refactor `backend-flask/app.py
+### Refactor `backend-flask/app.py`
 - Made substantial changes to `backend-flask\app.py`.
+
+### Refactor Flask Routes
+- Created `routes` directory.
